third_party/WebKit/LayoutTests/fast/css/invalidation/slotted-expected.txt

Style invalidation of slotted elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(slotted).backgroundColor is "rgba(0, 0, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(slotted).backgroundColor is "rgb(255, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(slotted).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
